refactor(core): extract pipeline frame, position, and layout helpers
Split post-bind stats/frame construction, data-space position adjust,
and layout/scene helper functions out of the pipeline orchestrator.
runPipeline is now primarily orchestration over focused modules.
Characterization tests pin buildFrame stats contracts and stack/dodge/jitter.
