+++ packages/core/src/pipeline.ts
@@ -35,66 +35,31 @@
  * zero-variance domain -> symmetric padding; log scales drop non-positive
  * values with a warning and REFUSE non-positive explicit domains.
  */
-import type {
-  BarParams,
-  BoxplotParams,
-  ErrorbarParams,
-  PortableSpec,
-  PositionParams,
-  PositionScaleSpec,
-  SmoothParams,
-  SpecInput,
-} from "@ggsvelte/spec";
+import type { BarParams, PortableSpec, SpecInput } from "@ggsvelte/spec";
 import { normalize, SpecValidationError, validate } from "@ggsvelte/spec";
 
-import { deriveGroups } from "./grouping.js";
 import { FONT_METRICS } from "./layout/font-metrics.js";
-import type { Domain, LayoutResult, Margins, PassResult, TickFormatter } from "./layout/layout.js";
+import type { LayoutResult, Margins, PassResult } from "./layout/layout.js";
 import { DEFAULT_LAYOUT_THEME, layout, layoutPass } from "./layout/layout.js";
 import { MetricsTableMeasurer } from "./layout/measure.js";
-import { formatTime, numberFormatter } from "./layout/format.js";
-import { defaultTickFormat, tickStep } from "./layout/ticks.js";
-import { defaultLogTickFormat } from "./layout/ticks.js";
-import { defaultTimeTickFormat } from "./layout/time.js";
 import type { LegendInput, LegendOrder } from "./legend.js";
 import { buildLegends } from "./legend.js";
 import type { ScaleState } from "./scales/state.js";
 import type { PositionScale } from "./scales/train.js";
 import { bandKey, finiteExtent } from "./scales/train.js";
-import type {
-  GeometryBatch,
-  Scene,
-  SceneAxis,
-  SceneLegend,
-  ScenePanel,
-  SceneTick,
-} from "./scene.js";
+import type { GeometryBatch, Scene, SceneAxis, SceneLegend, ScenePanel } from "./scene.js";
 import { PANEL_SPACING, STRIP_BAND } from "./scene.js";
-import { statBin } from "./stats/bin.js";
-import { statBoxplot } from "./stats/boxplot.js";
-import { statCount } from "./stats/count.js";
-import { statDensity } from "./stats/density.js";
-import { statSmooth } from "./stats/smooth.js";
-import { statSummary } from "./stats/summary.js";
-import { DEFAULT_JITTER_SEED, jitterOffsets, nudgeOffsets } from "./positions/jitter.js";
-import { positionDodge, positionStack } from "./positions/positions.js";
-import type { CellValue, Discreteness } from "./table.js";
-import { cellsToNumeric, cellToNumber, ColumnTable } from "./table.js";
+import type { CellValue } from "./table.js";
+import { cellToNumber, ColumnTable } from "./table.js";
 import { resolveEditionDefaults } from "./editions.js";
 import type { ThemeTokens } from "./theme.js";
 import { resolveTheme, UnknownThemeError } from "./theme.js";
 import { perfMark, perfMeasure } from "./perf.js";
 import { buildCandidateStore } from "./candidate-store.js";
-import type {
-  CandidateBuildFacts,
-  CandidateDatum,
-  ResolvedCandidateInspectMode,
-} from "./candidate-store.js";
 import { LineageStore } from "./identity.js";
 
 import type {
   Advisory,
-  AxisValueFormatter,
   LayerBackend,
   LayerBinding,
   LayerFrame,
@@ -110,12 +75,31 @@ import { batchMarkCount, buildBatch, flipBatchInPlace } from "./pipeline/geometr
 import type { Frame } from "./pipeline/geometry.js";
 import { resolveFacet, SINGLE_PANEL } from "./pipeline/facets.js";
 import { bindData, bindLayer } from "./pipeline/bind.js";
+import { collectAxisInputs, resolveColorScale, trainAxis } from "./pipeline/scale-training.js";
 import {
-  collectAxisInputs,
-  isBarLike,
-  resolveColorScale,
-  trainAxis,
-} from "./pipeline/scale-training.js";
+  buildFrame,
+  createRawCandidateDatumResolver,
+  candidateAutoMode,
+  deriveLayerGroups,
+  remapSourceRows,
+} from "./pipeline/frame.js";
+import { applyPosition } from "./pipeline/position.js";
+import {
+  AXIS_TITLE_BAND,
+  CAPTION_BAND,
+  LEGEND_EDGE_PAD,
+  LEGEND_GAP,
+  SUBTITLE_BAND,
+  TITLE_BAND,
+  axisTicks,
+  dedupeAdvisories,
+  dedupeWarnings,
+  elementwiseMaxMargins,
+  layoutDomain,
+  makeAxisFormatter,
+  makeAxisValueFormatter,
+  scaleDomainSnapshot,
+} from "./pipeline/layout-helpers.js";
 
 // Re-export the public pipeline contract (import path stability).
 export type {
@@ -133,679 +117,6 @@ export type {
 } from "./pipeline/types.js";
 export { CANVAS_AUTO_THRESHOLD, PipelineError } from "./pipeline/types.js";
 export { batchMarkCount } from "./pipeline/geometry.js";
-
-// ---------------------------------------------------------------------------
-// Layer frames (post-stat, post-position data)
-// ---------------------------------------------------------------------------
-
-/** Fresh all-null frame extras (each stat branch fills what it uses). */
-function emptyFrameExtras(): Pick<
-  LayerFrame,
-  | "ymin"
-  | "ymax"
-  | "xmin"
-  | "xmax"
-  | "dodgeSlot"
-  | "dodgeSlotCounts"
-  | "offsetX"
-  | "offsetY"
-  | "box"
-  | "smoothBand"
-  | "xIntercepts"
-  | "yIntercepts"
-> {
-  return {
-    ymin: null,
-    ymax: null,
-    xmin: null,
-    xmax: null,
-    dodgeSlot: null,
-    dodgeSlotCounts: null,
-    offsetX: null,
-    offsetY: null,
-    box: null,
-    smoothBand: false,
-    xIntercepts: [],
-    yIntercepts: [],
-  };
-}
-
-function interceptList(value: unknown): CellValue[] {
-  if (value === undefined) return [];
-  if (Array.isArray(value)) return value as CellValue[];
-  return [value as CellValue];
-}
-
-function deriveLayerGroups(binding: LayerBinding, table: ColumnTable): number[] {
-  const aes = binding.layer.aes ?? {};
-  const declared: Record<string, Discreteness> = {};
-  for (const mapping of Object.values(aes)) {
-    if (
-      mapping !== null &&
-      mapping !== undefined &&
-      "field" in mapping &&
-      table.has(mapping.field)
-    ) {
-      declared[mapping.field] = table.discreteness(mapping.field);
-    }
-  }
-  return [...deriveGroups(table.columns(), aes, declared).groups];
-}
-
-function createRawCandidateDatumResolver(
-  bindings: readonly LayerBinding[],
-  table: ColumnTable,
-  color: ResolvedColorScale | null,
-  fill: ResolvedColorScale | null,
-  lineage: LineageStore<number>,
-): (facts: CandidateBuildFacts) => CandidateDatum {
-  const groupsByLayer = new Map<number, readonly number[]>();
-  const groupsFor = (layerIndex: number): readonly number[] => {
-    let groups = groupsByLayer.get(layerIndex);
-    if (groups === undefined) {
-      const binding = bindings[layerIndex];
-      groups = binding === undefined ? [] : deriveLayerGroups(binding, table);
-      groupsByLayer.set(layerIndex, groups);
-    }
-    return groups;
-  };
-  return (facts) => {
-    const binding = bindings[facts.layerIndex];
-    const sourceRow = facts.rowIndex;
-    if (binding === undefined || sourceRow === null) return {};
-    const value = (field: string | null): CellValue =>
-      field === null ? null : table.column(field)[sourceRow]!;
-    const group = groupsFor(facts.layerIndex)[sourceRow] ?? 0;
-    const ordinalRank = (resolved: ResolvedColorScale | null, field: string | null) => {
-      if (resolved?.kind !== "ordinal" || field === null) return -1;
-      const key = bandKey(value(field));
-      return resolved.scale.domain.findIndex((domainValue) => bandKey(domainValue) === key);
-    };
-    const colorRank = ordinalRank(color, binding.color.field);
-    const fillRank = ordinalRank(fill, binding.fill.field);
-    return {
-      xValue: value(binding.xField),
-      yValue: value(binding.yField),
-      seriesId: group,
-      seriesRank: colorRank >= 0 ? colorRank : fillRank >= 0 ? fillRank : group,
-      sourceOrder: sourceRow,
-      lineage: lineage.intern([sourceRow]),
-      autoMode: candidateAutoMode(binding, facts.primitiveIndex),
-    };
-  };
-}
-
-function candidateAutoMode(
-  binding: LayerBinding,
-  primitiveIndex: number,
-): ResolvedCandidateInspectMode {
-  switch (binding.layer.geom) {
-    case "point":
-    case "text":
-      return "xy";
-    case "col":
-    case "bar":
-      return "exact";
-    case "line":
-    case "area":
-    case "density":
-    case "smooth":
-    case "errorbar":
-    case "boxplot":
-      return "x";
-    case "rule": {
-      if (binding.ruleForm === "vertical") return "x";
-      if (binding.ruleForm === "horizontal") return "y";
-      const params = (binding.layer.params ?? {}) as { xintercept?: unknown };
-      return primitiveIndex < interceptList(params.xintercept).length ? "x" : "y";
-    }
-    default:
-      return "xy";
-  }
-}
-
-/** Carried discrete columns for stats (color/fill/label, minus the x field). */
-function carriedColumns(
-  binding: LayerBinding,
-  table: ColumnTable,
-): Record<string, readonly CellValue[]> {
-  const carried: Record<string, readonly CellValue[]> = {};
-  for (const field of [binding.color.field, binding.fill.field, binding.labelField]) {
-    if (field !== null && field !== binding.xField) carried[field] = table.column(field);
-  }
-  return carried;
-}
-
-function removedStatWarning(
-  dropped: number,
-  index: number,
-  what: string,
-  warnings: PipelineWarning[],
-): void {
-  if (dropped > 0) {
-    warnings.push({
-      code: "removed-missing",
-      message: `Removed ${dropped} row(s) with ${what} (layer ${index}).`,
-    });
-  }
-}
-
-function buildFrame(
-  binding: LayerBinding,
-  table: ColumnTable,
-  warnings: PipelineWarning[],
-  advisories: Advisory[],
-  binRange?: [number, number],
-): LayerFrame {
-  const { layer, index } = binding;
-  const n = table.rowCount;
-
-  if (binding.ruleForm === "annotation") {
-    const params = (layer.params ?? {}) as { xintercept?: unknown; yintercept?: unknown };
-    return {
-      binding,
-      table,
-      n: 0,
-      xValues: null,
-      xNumeric: null,
-      yNumeric: null,
-      groups: [],
-      rowIndex: new Uint32Array(0),
-      colorValues: null,
-      fillValues: null,
-      labelValues: null,
-      ...emptyFrameExtras(),
-      xIntercepts: interceptList(params.xintercept),
-      yIntercepts: interceptList(params.yintercept),
-    };
-  }
-
-  const groups = deriveLayerGroups(binding, table);
-  const stat = layer.stat ?? "identity";
-  const carried = carriedColumns(binding, table);
-  const columnOf =
-    (result: { carried: Record<string, CellValue[]> }, x: readonly CellValue[] | null) =>
-    (field: string | null): readonly CellValue[] | null =>
-      field === null ? null : field === binding.xField ? x : (result.carried[field] ?? null);
-
-  if (stat === "count") {
-    const xField = binding.xField!;
-    const result = statCount({
-      x: table.column(xField),
-      groups,
-      weights: binding.weightField === null ? null : table.numeric(binding.weightField),
-      carried,
-    });
-    removedStatWarning(
-      result.dropped,
-      index,
-      "missing x or non-finite weight before counting",
-      warnings,
-    );
-    const col = columnOf(result, result.x);
-    return {
-      binding,
-      table,
-      n: result.x.length,
-      xValues: result.x,
-      xNumeric: cellsToNumeric(result.x),
-      yNumeric: binding.yStatColumn === "count" ? result.count : null,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-    };
-  }
-
-  if (stat === "bin") {
-    const result = statBin({
-      x: table.numeric(binding.xField!),
-      groups,
-      weights: binding.weightField === null ? null : table.numeric(binding.weightField),
-      carried,
-      params: (layer.params ?? {}) as BarParams,
-      // Fixed-x facets share one break grid across panels (ggplot2 derives
-      // breaks from the shared scale dimension); free_x omits this.
-      ...(binRange !== undefined && { range: binRange }),
-    });
-    removedStatWarning(result.dropped, index, "missing or non-finite x before binning", warnings);
-    if (result.usedDefaultBins && result.x.length > 0) {
-      // Hadley lesson 12 — the model advisory: geom_histogram's
-      // "`stat_bin()` using `bins = 30`. Pick better value with `binwidth`."
-      advisories.push({
-        code: "bin-default-bins",
-        path: `layers.${index}`,
-        chosen: "stat bin using bins = 30",
-        howToOverride: `Set params.binwidth (preferred — a width meaningful for the data) or params.bins on layer ${index}.`,
-      });
-    }
-    const columns: Record<string, Float64Array> = {
-      count: result.count,
-      density: result.density,
-      ncount: result.ncount,
-      ndensity: result.ndensity,
-    };
-    const col = columnOf(result, null);
-    return {
-      binding,
-      table,
-      n: result.x.length,
-      xValues: null,
-      xNumeric: result.x,
-      yNumeric: columns[binding.yStatColumn ?? "count"] ?? result.count,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-      xmin: result.xmin,
-      xmax: result.xmax,
-    };
-  }
-
-  if (stat === "density") {
-    const result = statDensity({
-      x: table.numeric(binding.xField!),
-      groups,
-      weights: binding.weightField === null ? null : table.numeric(binding.weightField),
-      carried,
-      params: (layer.params ?? {}) as { bw?: number; adjust?: number; n?: number; cut?: number },
-    });
-    removedStatWarning(result.dropped, index, "missing or non-finite x", warnings);
-    if (result.droppedGroups > 0) {
-      warnings.push({
-        code: "density-group-dropped",
-        message: `Layer ${index} (density): ${result.droppedGroups} group(s) with fewer than two data points have been dropped.`,
-      });
-    }
-    const columns: Record<string, Float64Array> = {
-      density: result.density,
-      count: result.count,
-      scaled: result.scaled,
-      ndensity: result.ndensity,
-    };
-    const yNumeric = columns[binding.yStatColumn ?? "density"] ?? result.density;
-    const col = columnOf(result, null);
-    const outN = result.x.length;
-    return {
-      binding,
-      table,
-      n: outN,
-      xValues: null,
-      xNumeric: result.x,
-      yNumeric,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: outN }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-      // Density renders as an area from the zero baseline.
-      ymin: new Float64Array(outN),
-      ymax: yNumeric,
-    };
-  }
-
-  if (stat === "smooth") {
-    const params = (layer.params ?? {}) as SmoothParams;
-    const result = statSmooth({
-      x: table.numeric(binding.xField!),
-      y: table.numeric(binding.yField!),
-      groups,
-      carried,
-      params,
-    });
-    removedStatWarning(result.dropped, index, "missing or non-finite x/y", warnings);
-    if (result.droppedGroups > 0) {
-      warnings.push({
-        code: "smooth-group-dropped",
-        message: `Layer ${index} (smooth): ${result.droppedGroups} group(s) too small or degenerate to fit have been dropped.`,
-      });
-    }
-    if (result.methodInferred && result.x.length > 0) {
-      advisories.push({
-        code: "smooth-method-inferred",
-        path: `layers.${index}`,
-        chosen: `stat smooth using method = "${result.methodUsed}" (largest group ${result.methodUsed === "loess" ? "<" : ">="} 1000 rows; ggplot2 would escalate to gam, which ggsvelte does not ship)`,
-        howToOverride: `Set params.method ("lm" | "loess") on layer ${index}.`,
-      });
-    }
-    const col = columnOf(result, null);
-    return {
-      binding,
-      table,
-      n: result.x.length,
-      xValues: null,
-      xNumeric: result.x,
-      yNumeric: result.y,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-      ymin: result.ymin,
-      ymax: result.ymax,
-      smoothBand: result.hasBand,
-    };
-  }
-
-  if (stat === "boxplot") {
-    const params = (layer.params ?? {}) as BoxplotParams;
-    const result = statBoxplot({
-      x: table.column(binding.xField!),
-      y: table.numeric(binding.yField!),
-      groups,
-      ...(params.coef !== undefined && { coef: params.coef }),
-      carried,
-    });
-    removedStatWarning(result.dropped, index, "missing or non-finite y", warnings);
-    const col = columnOf(result, result.x);
-    return {
-      binding,
-      table,
-      n: result.x.length,
-      xValues: result.x,
-      xNumeric: null,
-      yNumeric: null,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-      ymin: result.ymin,
-      ymax: result.ymax,
-      box: {
-        lower: result.lower,
-        middle: result.middle,
-        upper: result.upper,
-        outlierX: result.outliers.map((o) => o.x),
-        outlierY: Float64Array.from(result.outliers.map((o) => o.y)),
-        outlierBox: Uint32Array.from(result.outliers.map((o) => o.boxRow)),
-        outlierRow: Uint32Array.from(result.outliers.map((o) => o.sourceRow)),
-      },
-    };
-  }
-
-  if (stat === "summary") {
-    const params = (layer.params ?? {}) as ErrorbarParams;
-    const result = statSummary({
-      x: table.column(binding.xField!),
-      y: table.numeric(binding.yField!),
-      groups,
-      fun: params.fun,
-      funMin: params.funMin,
-      funMax: params.funMax,
-      carried,
-    });
-    removedStatWarning(result.dropped, index, "missing x or non-finite y", warnings);
-    const col = columnOf(result, result.x);
-    return {
-      binding,
-      table,
-      n: result.x.length,
-      xValues: result.x,
-      xNumeric: cellsToNumeric(result.x),
-      yNumeric: result.y,
-      groups: result.groups,
-      rowIndex: Uint32Array.from({ length: result.x.length }, () => NO_ROW),
-      colorValues: col(binding.color.field),
-      fillValues: col(binding.fill.field),
-      labelValues: col(binding.labelField),
-      ...emptyFrameExtras(),
-      ymin: result.ymin,
-      ymax: result.ymax,
-    };
-  }
-
-  // --- identity ----------------------------------------------------------------
-  return {
-    binding,
-    table,
-    n,
-    xValues: binding.xField === null ? null : table.column(binding.xField),
-    xNumeric: binding.xField === null ? null : table.numeric(binding.xField),
-    yNumeric: binding.yField === null ? null : table.numeric(binding.yField),
-    groups,
-    rowIndex: Uint32Array.from({ length: n }, (_, i) => i),
-    colorValues: binding.color.field === null ? null : table.column(binding.color.field),
-    fillValues: binding.fill.field === null ? null : table.column(binding.fill.field),
-    labelValues: binding.labelField === null ? null : table.column(binding.labelField),
-    ...emptyFrameExtras(),
-    ymin: binding.yminField === null ? null : table.numeric(binding.yminField),
-    ymax: binding.ymaxField === null ? null : table.numeric(binding.ymaxField),
-  };
-}
-
-/** Per-row slot keys: band categories, or bin centers for binned bars. */
-function slotKeys(frame: LayerFrame): (number | string)[] | null {
-  if (frame.xValues !== null) return frame.xValues.map((v) => bandKey(v));
-  if (frame.xNumeric !== null) return Array.from(frame.xNumeric);
-  return null;
-}
-
-function applyPosition(frame: LayerFrame, advisories: Advisory[], table: ColumnTable): void {
-  const { binding } = frame;
-  const layer = binding.layer;
-  const geom = layer.geom;
-
-  // --- jitter / nudge (point + text layers) -----------------------------------
-  if (geom === "point" || geom === "text") {
-    const position = layer.position ?? "identity";
-    if (position === "identity") return;
-    const params: PositionParams = "positionParams" in layer ? (layer.positionParams ?? {}) : {};
-    // Offsets are band-step fractions on discrete axes (resolution 1),
-    // data units on continuous axes.
-    const xDiscrete = binding.xField !== null && table.discreteness(binding.xField) === "discrete";
-    const yDiscrete = binding.yField !== null && table.discreteness(binding.yField) === "discrete";
-    if (position === "nudge") {
-      const { dx, dy } = nudgeOffsets(frame.n, params.x ?? 0, params.y ?? 0);
-      frame.offsetX = dx;
-      frame.offsetY = dy;
-      return;
-    }
-    // jitter (point only, schema-enforced): seeded — deliberate divergence
-    // from ggplot2's random jitter (decision 0010), always surfaced.
-    const { dx, dy } = jitterOffsets({
-      n: frame.n,
-      width: params.width,
-      height: params.height,
-      seed: params.seed,
-      xNumeric: xDiscrete ? null : frame.xNumeric,
-      yNumeric: yDiscrete ? null : frame.yNumeric,
-    });
-    frame.offsetX = dx;
-    frame.offsetY = dy;
-    advisories.push({
-      code: "jitter-seeded",
-      path: `layers.${binding.index}`,
-      chosen: `deterministic seeded jitter (seed ${params.seed ?? DEFAULT_JITTER_SEED}) — ggplot2 draws new random offsets every render; ggsvelte seeds for reproducibility`,
-      howToOverride: `Set positionParams.seed (and width/height) on layer ${binding.index}.`,
-    });
-    return;
-  }
-
-  // --- boxplot dodge -----------------------------------------------------------
-  if (geom === "boxplot") {
-    if ((layer.position ?? "dodge") !== "dodge" || frame.xValues === null) return;
-    const dodge = positionDodge({
-      slots: frame.xValues.map((v) => bandKey(v)),
-      groups: frame.groups,
-    });
-    frame.dodgeSlot = dodge.slot;
-    frame.dodgeSlotCounts = dodge.slotCount;
-    return;
-  }
-
-  // --- bar-like stack/fill/dodge ------------------------------------------------
-  if (!isBarLike(geom) || frame.yNumeric === null) return;
-  const slots = slotKeys(frame);
-  if (slots === null) return;
-  const position = layer.position ?? "identity";
-  const y = frame.yNumeric;
-
-  if (position === "stack" || position === "fill") {
-    const { ymin, ymax } = positionStack({ slots, groups: frame.groups, y, mode: position });
-    frame.ymin = ymin;
-    frame.ymax = ymax;
-    return;
-  }
-  // identity / dodge: bars grow from the zero baseline.
-  const ymin = new Float64Array(frame.n);
-  const ymax = new Float64Array(frame.n);
-  for (let i = 0; i < frame.n; i++) {
-    const v = Number.isFinite(y[i]!) ? y[i]! : 0;
-    ymin[i] = Math.min(0, v);
-    ymax[i] = Math.max(0, v);
-  }
-  frame.ymin = ymin;
-  frame.ymax = ymax;
-  if (position === "dodge") {
-    const dodge = positionDodge({ slots, groups: frame.groups });
-    frame.dodgeSlot = dodge.slot;
-    frame.dodgeSlotCounts = dodge.slotCount;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Layout + scene assembly
-// ---------------------------------------------------------------------------
-
-const TITLE_BAND = 22;
-const SUBTITLE_BAND = 16;
-const CAPTION_BAND = 14;
-const AXIS_TITLE_BAND = 18;
-const LEGEND_GAP = 12;
-const LEGEND_EDGE_PAD = 2;
-
-function layoutDomain(
-  scale: PositionScale,
-  breaks: readonly (number | string)[] | undefined,
-): Domain {
-  if (scale.type === "band") return { type: "band", categories: [...scale.domain] };
-  const numericBreaks =
-    breaks === undefined
-      ? undefined
-      : breaks.map((b) => cellToNumber(b as CellValue)).filter((v) => Number.isFinite(v));
-  return {
-    type: scale.type,
-    min: scale.domain[0],
-    max: scale.domain[1],
-    ...(numericBreaks !== undefined && { breaks: numericBreaks }),
-  };
-}
-
-function makeAxisFormatter(
-  axis: "x" | "y",
-  scale: PositionScale,
-  config: PositionScaleSpec | undefined,
-  warnings: PipelineWarning[],
-): TickFormatter | undefined {
-  const labels = config?.labels;
-  if (labels === undefined) return undefined;
-  if (scale.type === "band") {
-    warnings.push({
-      code: "invalid-label-format",
-      message: `scales.${axis}.labels format strings apply to continuous scales; the band ${axis} scale ignores it.`,
-    });
-    return undefined;
-  }
-  if (scale.type === "time") {
-    return (value) => formatTime(value as number, labels);
-  }
-  const f = numberFormatter(labels);
-  if (!f.ok) {
-    warnings.push({
-      code: "invalid-label-format",
-      message: `Unrecognized labels format "${labels}" on scales.${axis}; using the default.`,
-    });
-    return undefined;
-  }
-  return (value) => f.format(value as number);
-}
-
-function makeAxisValueFormatter(
-  scale: PositionScale,
-  custom: TickFormatter | undefined,
-): AxisValueFormatter {
-  if (scale.type === "band") return (value) => (value === null ? "–" : String(value));
-  const fallback =
-    scale.type === "time"
-      ? defaultTimeTickFormat
-      : scale.type === "log"
-        ? defaultLogTickFormat
-        : defaultTickFormat(tickStep(scale.domain[0], scale.domain[1], 5));
-  return (value) => {
-    if (value === null) return "–";
-    const numeric = cellToNumber(value);
-    return custom === undefined ? fallback(numeric) : custom(numeric, NaN);
-  };
-}
-
-/** Project layout ticks through a scale onto an axis extent, in px. */
-function axisTicks(
-  scale: PositionScale,
-  ticks: LayoutResult["x"]["ticks"],
-  extent: number,
-  fromEnd: boolean,
-): SceneTick[] {
-  const out: SceneTick[] = [];
-  for (const tick of ticks) {
-    const t =
-      scale.type === "band" ? scale.normalize(tick.value) : scale.normalize(tick.value as number);
-    if (t === undefined || Number.isNaN(t)) continue;
-    const pos = fromEnd ? extent - t * extent : t * extent;
-    out.push({ pos, label: tick.labeled ? tick.label : "" });
-  }
-  return out;
-}
-
-function dedupeWarnings(list: PipelineWarning[]): PipelineWarning[] {
-  const seen = new Set<string>();
-  return list.filter((w) => {
-    const key = `${w.code} ${w.message}`;
-    if (seen.has(key)) return false;
-    seen.add(key);
-    return true;
-  });
-}
-
-function dedupeAdvisories(list: Advisory[]): Advisory[] {
-  const seen = new Set<string>();
-  return list.filter((a) => {
-    const key = `${a.code} ${a.path}`;
-    if (seen.has(key)) return false;
-    seen.add(key);
-    return true;
-  });
-}
-
-const elementwiseMaxMargins = (a: Margins, b: Margins): Margins => ({
-  top: Math.max(a.top, b.top),
-  right: Math.max(a.right, b.right),
-  bottom: Math.max(a.bottom, b.bottom),
-  left: Math.max(a.left, b.left),
-});
-
-/**
- * Facet frames index into the PANEL table; hit-testing/tooltips need SOURCE
- * rows, so remap through the partition (NO_ROW stays NO_ROW).
- */
-function remapSourceRows(frame: LayerFrame, sourceRows: number[] | null): void {
-  if (sourceRows === null) return;
-  for (let i = 0; i < frame.rowIndex.length; i++) {
-    const local = frame.rowIndex[i]!;
-    if (local !== NO_ROW) frame.rowIndex[i] = sourceRows[local]!;
-  }
-}
-
-function scaleDomainSnapshot(scale: PositionScale): readonly CellValue[] {
-  return Object.freeze([...scale.domain]);
-}
 
 // ---------------------------------------------------------------------------
 // runPipeline
